reorganization_backup_root/requirements_enhanced.txt
# Enhanced GraphRAG MCP Server Requirements
# ==========================================

# Core MCP server dependencies
asyncio-mqtt>=0.13.0
websockets>=11.0.3
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# IPFS Kit dependencies
requests>=2.31.0
httpx>=0.24.0                # For async HTTP requests
aiohttp>=3.9.0
ipfshttpclient>=0.8.0

# LibP2P peer management dependencies
libp2p>=0.2.8               # Core libp2p functionality
multiaddr>=0.0.9            # For peer addressing
multiformats>=0.2.0         # For content addressing
base58>=2.1.1               # Used by CIDs and peer IDs
cryptography>=38.0.0        # For key generation and encryption
protobuf>=3.20.0,<5.0.0     # Compatible protobuf version range

# Search and AI dependencies (optional but recommended)
sentence-transformers>=2.2.0  # For vector embeddings
scikit-learn>=1.3.0          # For similarity calculations
numpy>=1.24.0                # For numerical operations
networkx>=3.0                # For knowledge graphs
rdflib>=7.0.0                # For SPARQL queries

# Additional ML/AI dependencies (optional)
transformers>=4.35.0         # For advanced NLP
torch>=2.0.0                # For deep learning models
faiss-cpu>=1.7.4            # For efficient vector search

# Data processing dependencies
pandas>=2.0.0               # For data manipulation
jsonlines>=4.0.0           # For streaming JSON data

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0

# Optional GPU acceleration (uncomment if you have CUDA)
# torch>=2.0.0+cu118
# faiss-gpu>=1.7.4
